# api/src/openapi.yaml
openapi: 3.0.2
info:
  version: 0.1.0
  title: NFT Gallery API
  description: API serving the NFT Gallery Platform
...
security:
  - bearerAuth: []
servers:
  - url: '{protocol}://{host}:{port}'
    description: 'Local Mocks (e.g. http://localhost:4010) and Local Dev Server (e.g. http://localhost:3000/v1)'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
        description: Protocol of the URL.
      host:
        enum:
          - api-dev.domi.house
          - api.domi.house
          - localhost
        default: localhost
        description: Host of the URL
      port:
        enum:
          - ''
          - '4010'
          - '3000'
        default: '3000'
        description: Port of the URL
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication token
paths:
...
  /v1/collection:
    post:
      summary: Create Collection
      operationId: createCollection
      x-eov-operation-handler: CollectionsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Collection
      requestBody:
        description: information that is needed for Collection
        required: true
        content:
          application/json:
            schema:
              title: Collection
              type: object
              description: Nft Gallery Collection of Assets
              x-autogen:
                openapi: true
                ownership: true
                mongo: true
                security:
                  create:
                    - admin
                    - consumer
                  update:
                    - admin
                    - consumer
                  delete:
                    - admin
                    - consumer
                  retrieve:
                    - admin
                    - consumer
                  retrieves:
                    - admin
                    - consumer
              properties:
                ownerId:
                  type: string
                title:
                  type: string
                owner:
                  type: string
                category:
                  type: string
                  enum:
                    - Art
                    - Photography
                    - Games
                    - Metaverses
                    - Music
                    - Domains
                    - Memes
              required:
                - title
                - owner
                - category
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
  /v1/collections:
    parameters:
      - name: limit
        in: query
        description: How many items to return at one time (max 100)
        required: false
        schema:
          type: integer
          format: int32
      - name: skip
        in: query
        description: How many items should be skipped before returnig the required amount
        required: false
        schema:
          type: integer
          format: int32
    get:
      summary: Get all Collections
      operationId: retrieveCollections
      x-eov-operation-handler: CollectionsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Collection
      parameters:
        - name: ownerId
          in: query
          required: false
          description: The Fileter ownerId
          schema:
            type: string
            example: null
        - name: title
          in: query
          required: false
          description: The Fileter title
          schema:
            type: string
            example: New Title
        - name: owner
          in: query
          required: false
          description: The Fileter owner
          schema:
            type: string
            example: null
        - name: category
          in: query
          required: false
          description: The Fileter category
          schema:
            type: string
            default: null
            enum:
              - Art
              - Photography
              - Games
              - Metaverses
              - Music
              - Domains
              - Memes
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
  '/v1/collections/{collectionId}':
    parameters:
      - name: collectionId
        in: path
        required: true
        description: The Id of the Collection
        schema:
          description: Id of the entity
          type: string
          format: not-empty-string
          example: 60758e80d59c357138d01af6
    get:
      summary: Get Collection with Id
      operationId: retrieveCollection
      x-eov-operation-handler: CollectionsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Collection
      parameters:
        - name: collectionId
          in: path
          required: true
          description: The Id of the Collection
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
    put:
      summary: Update Collection with Id
      operationId: updateCollection
      x-eov-operation-handler: CollectionsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Collection
      parameters:
        - name: collectionId
          in: path
          required: true
          description: The Id of the Collection
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      requestBody:
        description: information that is needed for Collection
        required: true
        content:
          application/json:
            schema:
              title: Update Collection with Id
              type: object
              description: An array of fields that we want to change name and value
              properties:
                ownerId:
                  type: string
                title:
                  type: string
                owner:
                  type: string
                category:
                  type: string
                  enum:
                    - Art
                    - Photography
                    - Games
                    - Metaverses
                    - Music
                    - Domains
                    - Memes
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
    delete:
      summary: Delete Collection with Id
      operationId: deleteCollection
      x-eov-operation-handler: CollectionsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Collection
      parameters:
        - name: collectionId
          in: path
          required: true
          description: The Id of the Collection
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
...